Fix embedded zip fs support on Linux
Either the Go linker is producing garbage at the end of executables on
Linux, or an extra section isn't being recorded in the ELF data on that
platform when cgo is used with a binary. Regardless, this issue made it
impossible to find the correct starting point within the binary for the
zip content. The zip reader, however, can locate the right data if we
first run the zip utility's "adjust self-extracting exe" fixup on our
binary and then just tell it the whole file is a zip. So... at the
expensive of having the zip reader traverse the whole binary looking for
the start of the zip data, we now support Linux even when cgo is
involved. (sigh)