Use generics for the sets
- These now support any comparable type.
- Added .Len() to the sets.
- Removed the json & yaml marshal/unmarshal support (which emitted the
  contents as an array), since there is no way to make them output in
  a predictable order, thanks to the comparable types only supporting
  == and !=. Could have constrained this to the contraints.Ordered
  types to make that work... but chose not to.

This is a breaking change for those using the old sets, partially
because of the removal of json & yaml support (note: you can still use
json & yaml marshaling, but the set will be emitted as a map, which is
the underlying type), but also because the old types no longer exist and
have been replaced with the generic syntax, i.e. Set[int], Set[string],
etc.